#9 Source code restructure.
