+++ index.js
@@ -1,1 +1,2 @@
+// @flow strict
 export { Model, Schema, Database } from "./src";
