+++ src/database/index.js
@@ -1,0 +1,4 @@
+// @flow strict
+export { Session } from "./Session";
+export type { Table, TableRowType } from "./Table";
+export type { Database, StateType, ActionType } from "./Database";
