+++ src/model/index.js
@@ -1,0 +1,3 @@
+// @flow strict
+export { Model } from "./Model";
+export type { Schema, ModelFieldsType } from "./Schema";
